refactor: update home page component file strucutre
