+++ src/components/(marketing)/home/MainDiscordEmbed.tsx
